requirements.txt
airtable==0.4.8
annotated-types==0.7.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
httplib2==0.22.0
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.0
MarkupSafe==3.0.2
numpy==2.3.1
pandas==2.3.0
praw==7.8.1
prawcore==2.4.0
proto-plus==1.26.1
protobuf==5.29.5
pyairtable==3.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytrends==4.9.2
pytz==2025.2
requests==2.32.4
rsa==4.9.1
six==1.17.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
update-checker==0.18.0
uritemplate==4.2.0
urllib3==2.5.0
websocket-client==1.8.0
Werkzeug==3.1.3
